# src/config/swagger.config.json
{
    "swagger": "2.0",
    "info": {
        "title": "Link Up API Documentation",
        "version": "1.0.0",
        "description": "API for the Link Up Social Media Application",
        "contact": {
            "name": "Paul Sanga Nyamawi",
            "email": "paulsanganyamawi@gmail.com"
        },
        "licence": {
            "name": "TheJituLicence",
            "url": "www.thejitu.com"
        }
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/link-up/v1",
    "schemes": ["http"],
    "paths": {
...
        "/users": {
            "get": {
                "tags": ["User Management"],
                "summary": "Fetch all users",
                "description": "It will return an array of user account details",
                "operationId": "fetchAllUsers",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Fetch successful"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": ["User Management"],
                "summary": "Fetch user by id",
                "description": "It will return a single user account details",
                "operationId": "FetchUserById",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "User's account id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetch successful"
                    },
                    "404": {
                        "description": "User account not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },

            "put": {
                "tags": ["User Management"],
                "summary": "Update user account details",
                "description": "User provides new account details and the old data is overwitten",
                "operationId": "updateUserAccount",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User's details to update",
                        "required": true,
                        "schema": {"$ref": "#/definitions/User"}
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update successful"
                    },
                    "400": {
                        "description": "Missing or empty request body"
                    },
                    "404": {
                        "description": "User account not found"
                    },
                    "422": {
                        "description": "User input validation failed"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },

            "delete": {
                "tags": ["User Management"],
                "summary": "Deletes user account",
                "description": "It deletes user account if it exits",
                "operationId": "deleteUserAccount",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "User account's id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletion successful"
                    },
                    "404": {
                        "description": "User account not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/post/{id}":{
            "get": {
                "tags": ["Post Management"],
                "summary": "Fetch a single post",
                "description": "It will return a post whose id matches the id on the request path",
                "operationId": "fetchPostById",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Post id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetch successful"
                    },
                    "404": {
                        "description": "post not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": ["Post Management"],
                "summary": "Create a new post",
                "description": "It will create a new post based on the request body",
                "operationId": "createNewPost",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Post owner id",
                        "required": true
                    }, 
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Post details to create",
                        "required": true,
                        "schema": {"$ref":"#/definitions/Post"}
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Post created successfully"
                    },
                    "422": {
                        "description": "Post input validation failed"
                    },
                    "400":{
                        "description": "Error creating post"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },

            "delete":{
                "tags": ["Post Management"],
                "summary": "Delete an existing post",
                "description": "It will delete an existing post",
                "operationId": "deletePost",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in":"path",
                        "name": "id",
                        "description": "Post id",
                        "required": true
                    }
                ],
                "responses":{
                    "200": {
                        "description": "Post deleted successfully"
                    },
                    "404": {
                        "description": "Post not found"
                    },
                    "409": {
                        "description": "Post already deleted"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "type": "object",
            "required": ["firstName", "lastName", "email", "profilePicture"],
            "properties": {
                "firstName": {
                    "type": "string",
                    "description": "User's first name",
                    "example": "Paul"
                },
                "lastName": {
                    "type": "string",
                    "description": "User's last name",
                    "example": "Sanga"
                },
                "email": {
                    "type": "string",
                    "description": "User's email address",
                    "example": "paul.nyamawi99@gmail.com"
                },
                "profilePicture": {
                    "type": "string",
                    "description": "User's profile picture",
                    "example": "https://www.google.com/imgres?imgurl=https%3A%2F%2Fwallpapercg.com%2Fdownload%2Fyuji-itadori-3840x2743-9543.jpg&tbnid=u9FKufbPMLQ2EM&vet=12ahUKEwibov3YmZOBAxXQkicCHat8CkAQMyhLegUIARClAg..i&imgrefurl=https%3A%2F%2Fwallpapercg.com%2Fyuji-itadori-wallpapers&docid=sk23cipVVPF6mM&w=3840&h=2743&q=itadori&ved=2ahUKEwibov3YmZOBAxXQkicCHat8CkAQMyhLegUIARClAg"
                }
            } 
        },
...
        "Post": {
            "type": "object",
            "required": ["image", "content"],
            "properties": {
                "image": {
                    "type": "string",
                    "description": "Post image",
                    "example": "https://pin.it/4SWOUSX"
                },
                "content": {
                    "type":"string",
                    "description": "Post content",
                    "example": "Brabus G-Wagon G900"
                }
            }
        }
    }
}